@@ -9,36 +9,33 @@
   },
   "private": false,
   "dependencies": {
-    "@emotion/react": "11.4.1",
-    "@emotion/styled": "11.3.0",
-    "@mui/material": "5.0.3",
-    "@mui/icons-material": "5.0.3",
-    "@mui/lab": "5.0.0-alpha.50",
-    "@mui/styles": "5.0.1",
-    "@types/nprogress": "0.2.0",
-    "@types/react": "17.0.25",
-    "@types/react-dom": "17.0.9",
-    "@types/react-router-dom": "5.3.0",
-    "chart.js": "2.9.4",
-    "clsx": "1.1.1",
-    "date-fns": "2.25.0",
-    "history": "5.0.1",
-    "nprogress": "0.2.0",
-    "numeral": "2.0.6",
-    "prop-types": "15.7.2",
-    "react": "17.0.2",
-    "react-custom-scrollbars-2": "4.4.0",
-    "react-dom": "17.0.2",
-    "react-helmet-async": "1.1.2",
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "@mui/icons-material": "^5.8.0",
+    "@mui/lab": "^5.0.0-alpha.82",
+    "@mui/material": "^5.8.0",
+    "@mui/styles": "^5.8.0",
+    "@types/nprogress": "^0.2.0",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.4",
+    "clsx": "^1.1.1",
+    "date-fns": "^2.28.0",
+    "history": "^5.3.0",
+    "nprogress": "^0.2.0",
+    "numeral": "^2.0.6",
+    "prop-types": "^15.8.1",
+    "react": "^18.1.0",
+    "react-custom-scrollbars-2": "^4.4.0",
+    "react-dom": "^18.1.0",
+    "react-helmet-async": "^1.3.0",
     "react-redux": "^8.0.1",
-    "react-router": "6.0.0-beta.0",
-    "react-router-dom": "6.0.0-beta.0",
-    "react-scripts": "^4.0.3",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
     "redux": "^4.2.0",
-    "stylis": "4.0.10",
-    "stylis-plugin-rtl": "2.1.0",
-    "react-chartjs-2": "2.11.1",
-    "web-vitals": "2.1.2"
+    "stylis": "^4.1.1",
+    "stylis-plugin-rtl": "^2.1.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "set PORT=3001 && react-scripts start",
@@ -62,13 +59,11 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
+  },"devDependencies": {
     "@babel/eslint-parser": "7.15.8",
     "@types/numeral": "2.0.2",
     "@types/react": "17.0.29",
     "@types/react-dom": "17.0.9",
-    "@types/react-router-dom": "5.3.1",
     "@typescript-eslint/eslint-plugin": "5.0.0",
     "@typescript-eslint/parser": "5.0.0",
     "eslint": "7.32.0",
