@@ -1,6 +1,6 @@
 import App from './App';
-import ReactDOM from 'react-dom';
-import 'src/utils/chart';
+
+
 import * as serviceWorker from './serviceWorker';
 import { HelmetProvider } from 'react-helmet-async';
 import { BrowserRouter } from 'react-router-dom';
@@ -9,9 +9,13 @@ import 'nprogress/nprogress.css';
 import { SidebarProvider } from './contexts/SidebarContext';
 import { createStore } from 'redux';
 import reducer from './store/reducer'
+import ReactDOM from "react-dom/client";
 
 const store = createStore(reducer);
-ReactDOM.render(
+
+const root = ReactDOM.createRoot(document.getElementById("root"));
+
+root.render(
   <Provider store={store}>
     <HelmetProvider>
       <SidebarProvider>
@@ -21,8 +25,6 @@ ReactDOM.render(
       </SidebarProvider>
     </HelmetProvider>
   </Provider>
-,
-  document.getElementById('root')
 );
 
 serviceWorker.unregister();
