@@ -1,6 +1,6 @@
 import { Suspense, lazy } from 'react';
 import { Navigate } from 'react-router-dom';
-import { PartialRouteObject } from 'react-router';
+import { RouteObject } from 'react-router';
 
 import SidebarLayout from 'src/layouts/SidebarLayout';
 import BaseLayout from 'src/layouts/BaseLayout';
@@ -49,13 +49,13 @@ const StatusComingSoon = Loader(lazy(() => import('src/content/pages/Status/Comi
 const StatusMaintenance = Loader(lazy(() => import('src/content/pages/Status/Maintenance')));
 
 
-const routes: PartialRouteObject[] = [
+const routes: RouteObject[] = [
   {
     path: '*',
     element: <BaseLayout />,
     children: [
       {
-        path: '/',
+        path: '',
         element: <Overview />
       },
       {
@@ -76,7 +76,7 @@ const routes: PartialRouteObject[] = [
         path: 'status',
         children: [
           {
-            path: '/',
+            path: '',
             element: (
               <Navigate
                 to="404"
@@ -115,7 +115,7 @@ const routes: PartialRouteObject[] = [
     ),
     children: [
       {
-        path: '/',
+        path: '',
         element: (
           <Navigate
             to="/dashboards/crypto"
@@ -140,7 +140,7 @@ const routes: PartialRouteObject[] = [
     ),
     children: [
       {
-        path: '/',
+        path: '',
         element: (
           <Navigate
             to="/management/transactions"
@@ -156,7 +156,7 @@ const routes: PartialRouteObject[] = [
         path: 'profile',
         children: [
           {
-            path: '/',
+            path: '',
             element: (
               <Navigate
                 to="details"
@@ -183,7 +183,7 @@ const routes: PartialRouteObject[] = [
     ),
     children: [
       {
-        path: '/',
+        path: '',
         element: (
           <Navigate
             to="/components/buttons"
